+++ database/migrations/2025_09_21_084420_create_resep_template_details_table.php
@@ -11,17 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('resep_template_details', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('template_id');
-            $table->string('kode_brng', 15);
-            $table->decimal('jumlah', 8, 2);
-            $table->text('aturan_pakai');
-            $table->timestamps();
+        if (!Schema::hasTable('resep_template_details')) {
+            Schema::create('resep_template_details', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('template_id');
+                $table->string('kode_brng', 15);
+                $table->decimal('jumlah', 8, 2);
+                $table->text('aturan_pakai');
+                $table->timestamps();
 
-            $table->foreign('template_id')->references('id')->on('resep_templates')->onDelete('cascade');
-            $table->index('kode_brng');
-        });
+                $table->foreign('template_id')->references('id')->on('resep_templates')->onDelete('cascade');
+                $table->index('kode_brng');
+            });
+        }
     }
 
     /**
